Refactoring of scenario atomics
- Scenario atomics were moved to srunner.scenariomanager.scenarioatomics
- Behavior atomics were split into behaviors and conditions

Along with that the imports in the scenarios were cleaned up
