+++ srunner/scenariomanager/scenarioatomics/atomic_behaviors.py
@@ -27,7 +27,7 @@ from agents.navigation.roaming_agent import *
 
 from srunner.scenariomanager.carla_data_provider import CarlaActorPool, CarlaDataProvider
 from srunner.scenariomanager.timer import GameTime
-from srunner.tools.scenario_helper import detect_lane_obstacle, get_distance_along_route
+from srunner.tools.scenario_helper import detect_lane_obstacle
 
 EPSILON = 0.001
 
@@ -82,367 +82,6 @@ class AtomicBehavior(py_trees.behaviour.Behaviour):
 
     def terminate(self, new_status):
         self.logger.debug("%s.terminate()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-
-class StandStill(AtomicBehavior):
-
-    """
-    This class contains a standstill behavior of a scenario
-    """
-
-    def __init__(self, actor, name, duration=float("inf")):
-        """
-        Setup actor
-        """
-        super(StandStill, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._actor = actor
-
-        self._duration = duration
-        self._start_time = 0
-
-    def initialise(self):
-        self._start_time = GameTime.get_time()
-        super(StandStill, self).initialise()
-
-    def update(self):
-        """
-        Check if the _actor stands still (v=0)
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        velocity = CarlaDataProvider.get_velocity(self._actor)
-
-        if velocity < EPSILON:
-            new_status = py_trees.common.Status.SUCCESS
-
-        if GameTime.get_time() - self._start_time > self._duration:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-
-class InTriggerRegion(AtomicBehavior):
-
-    """
-    This class contains the trigger region (condition) of a scenario
-    """
-
-    def __init__(self, actor, min_x, max_x, min_y, max_y, name="TriggerRegion"):
-        """
-        Setup trigger region (rectangle provided by
-        [min_x,min_y] and [max_x,max_y]
-        """
-        super(InTriggerRegion, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._actor = actor
-        self._min_x = min_x
-        self._max_x = max_x
-        self._min_y = min_y
-        self._max_y = max_y
-
-    def update(self):
-        """
-        Check if the _actor location is within trigger region
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        location = CarlaDataProvider.get_location(self._actor)
-
-        if location is None:
-            return new_status
-
-        not_in_region = (location.x < self._min_x or location.x > self._max_x) or (
-            location.y < self._min_y or location.y > self._max_y)
-        if not not_in_region:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-
-class InTriggerDistanceToVehicle(AtomicBehavior):
-
-    """
-    This class contains the trigger distance (condition) between to actors
-    of a scenario
-    """
-
-    def __init__(self, other_actor, actor, distance, name="TriggerDistanceToVehicle"):
-        """
-        Setup trigger distance
-        """
-        super(InTriggerDistanceToVehicle, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._other_actor = other_actor
-        self._actor = actor
-        self._distance = distance
-
-    def update(self):
-        """
-        Check if the ego vehicle is within trigger distance to other actor
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        ego_location = CarlaDataProvider.get_location(self._actor)
-        other_location = CarlaDataProvider.get_location(self._other_actor)
-
-        if ego_location is None or other_location is None:
-            return new_status
-
-        if calculate_distance(ego_location, other_location) < self._distance:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-
-class InTriggerDistanceToLocation(AtomicBehavior):
-
-    """
-    This class contains the trigger (condition) for a distance to a fixed
-    location of a scenario
-    """
-
-    def __init__(self, actor, target_location, distance, name="InTriggerDistanceToLocation"):
-        """
-        Setup trigger distance
-        """
-        super(InTriggerDistanceToLocation, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._target_location = target_location
-        self._actor = actor
-        self._distance = distance
-
-    def update(self):
-        """
-        Check if the actor is within trigger distance to the target location
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        location = CarlaDataProvider.get_location(self._actor)
-
-        if location is None:
-            return new_status
-
-        if calculate_distance(
-                location, self._target_location) < self._distance:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-
-class InTriggerDistanceToNextIntersection(AtomicBehavior):
-
-    """
-    This class contains the trigger (condition) for a distance to the
-    next intersection of a scenario
-    """
-
-    def __init__(self, actor, distance, name="InTriggerDistanceToNextIntersection"):
-        """
-        Setup trigger distance
-        """
-        super(InTriggerDistanceToNextIntersection, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._actor = actor
-        self._distance = distance
-        self._map = CarlaDataProvider.get_map()
-
-        waypoint = self._map.get_waypoint(self._actor.get_location())
-        while waypoint and not waypoint.is_intersection:
-            waypoint = waypoint.next(1)[-1]
-
-        self._final_location = waypoint.transform.location
-
-    def update(self):
-        """
-        Check if the actor is within trigger distance to the intersection
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        current_waypoint = self._map.get_waypoint(CarlaDataProvider.get_location(self._actor))
-        distance = calculate_distance(current_waypoint.transform.location, self._final_location)
-
-        if distance < self._distance:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-
-class TriggerVelocity(AtomicBehavior):
-
-    """
-    This class contains the trigger velocity (condition) of a scenario
-
-    The behavior is successful, if the actor is at least as fast as requested
-    """
-
-    def __init__(self, actor, target_velocity, name="TriggerVelocity"):
-        """
-        Setup trigger velocity
-        """
-        super(TriggerVelocity, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._actor = actor
-        self._target_velocity = target_velocity
-
-    def update(self):
-        """
-        Check if the actor has the trigger velocity
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        delta_velocity = self._target_velocity - CarlaDataProvider.get_velocity(self._actor)
-        if delta_velocity < EPSILON:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-
-class InTimeToArrivalToLocation(AtomicBehavior):
-
-    """
-    This class contains a check if a actor arrives within a given time
-    at a given location.
-    """
-
-    _max_time_to_arrival = float('inf')  # time to arrival in seconds
-
-    def __init__(self, actor, time, location, name="TimeToArrival"):
-        """
-        Setup parameters
-        """
-        super(InTimeToArrivalToLocation, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._actor = actor
-        self._time = time
-        self._target_location = location
-
-    def update(self):
-        """
-        Check if the actor can arrive at target_location within time
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        current_location = CarlaDataProvider.get_location(self._actor)
-
-        if current_location is None:
-            return new_status
-
-        distance = calculate_distance(current_location, self._target_location)
-        velocity = CarlaDataProvider.get_velocity(self._actor)
-
-        # if velocity is too small, simply use a large time to arrival
-        time_to_arrival = self._max_time_to_arrival
-        if velocity > EPSILON:
-            time_to_arrival = distance / velocity
-
-        if time_to_arrival < self._time:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-
-class InTimeToArrivalToVehicle(AtomicBehavior):
-
-    """
-    This class contains a check if a actor arrives within a given time
-    at another actor.
-    """
-
-    _max_time_to_arrival = float('inf')  # time to arrival in seconds
-
-    def __init__(self, other_actor, actor, time, name="TimeToArrival"):
-        """
-        Setup parameters
-        """
-        super(InTimeToArrivalToVehicle, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._other_actor = other_actor
-        self._actor = actor
-        self._time = time
-
-    def update(self):
-        """
-        Check if the ego vehicle can arrive at other actor within time
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        current_location = CarlaDataProvider.get_location(self._actor)
-        target_location = CarlaDataProvider.get_location(self._other_actor)
-
-        if current_location is None or target_location is None:
-            return new_status
-
-        distance = calculate_distance(current_location, target_location)
-        current_velocity = CarlaDataProvider.get_velocity(self._actor)
-        other_velocity = CarlaDataProvider.get_velocity(self._other_actor)
-
-        # if velocity is too small, simply use a large time to arrival
-        time_to_arrival = self._max_time_to_arrival
-
-        if current_velocity > other_velocity:
-            time_to_arrival = 2 * distance / (current_velocity - other_velocity)
-
-        if time_to_arrival < self._time:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-
-class InTriggerDistanceToLocationAlongRoute(AtomicBehavior):
-
-    """
-    Implementation for a behavior that will check if a given actor
-    is within a given distance to a given location considering a given route
-    """
-
-    def __init__(self, actor, route, location, distance, name="InTriggerDistanceToLocationAlongRoute"):
-        """
-        Setup class members
-        """
-        super(InTriggerDistanceToLocationAlongRoute, self).__init__(name)
-        self._map = CarlaDataProvider.get_map()
-        self._actor = actor
-        self._location = location
-        self._route = route
-        self._distance = distance
-
-        self._location_distance, _ = get_distance_along_route(self._route, self._location)
-
-    def update(self):
-        new_status = py_trees.common.Status.RUNNING
-
-        current_location = CarlaDataProvider.get_location(self._actor)
-
-        if current_location is None:
-            return new_status
-
-        if current_location.distance(self._location) < self._distance + 20:
-
-            actor_distance, _ = get_distance_along_route(self._route, current_location)
-
-            if (self._location_distance < actor_distance + self._distance and
-                actor_distance < self._location_distance) or \
-                    self._location_distance < 1.0:
-                new_status = py_trees.common.Status.SUCCESS
-
-        return new_status
 
 
 class AccelerateToVelocity(AtomicBehavior):
@@ -575,44 +214,6 @@ class KeepVelocity(AtomicBehavior):
         super(KeepVelocity, self).terminate(new_status)
 
 
-class DriveDistance(AtomicBehavior):
-
-    """
-    This class contains an atomic behavior to drive a certain distance.
-    """
-
-    def __init__(self, actor, distance, name="DriveDistance"):
-        """
-        Setup parameters
-        """
-        super(DriveDistance, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._target_distance = distance
-        self._distance = 0
-        self._location = None
-        self._actor = actor
-
-    def initialise(self):
-        self._location = CarlaDataProvider.get_location(self._actor)
-        super(DriveDistance, self).initialise()
-
-    def update(self):
-        """
-        Check driven distance
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        new_location = CarlaDataProvider.get_location(self._actor)
-        self._distance += calculate_distance(self._location, new_location)
-        self._location = new_location
-
-        if self._distance > self._target_distance:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-        return new_status
-
-
 class UseAutoPilot(AtomicBehavior):
 
     """
@@ -720,42 +321,6 @@ class StopVehicle(AtomicBehavior):
         self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
 
         return new_status
-
-
-class WaitForTrafficLightState(AtomicBehavior):
-
-    """
-    This class contains an atomic behavior to wait for a given traffic light
-    to have the desired state.
-    """
-
-    def __init__(self, traffic_light, state, name="WaitForTrafficLightState"):
-        """
-        Setup traffic_light
-        """
-        super(WaitForTrafficLightState, self).__init__(name)
-        self.logger.debug("%s.__init__()" % (self.__class__.__name__))
-        self._traffic_light = traffic_light
-        self._traffic_light_state = state
-
-    def update(self):
-        """
-        Set status to SUCCESS, when traffic light state is RED
-        """
-        new_status = py_trees.common.Status.RUNNING
-
-        # the next line may throw, if self._traffic_light is not a traffic
-        # light, but another actor. This is intended.
-        if str(self._traffic_light.state) == self._traffic_light_state:
-            new_status = py_trees.common.Status.SUCCESS
-
-        self.logger.debug("%s.update()[%s->%s]" % (self.__class__.__name__, self.status, new_status))
-
-        return new_status
-
-    def terminate(self, new_status):
-        self._traffic_light = None
-        super(WaitForTrafficLightState, self).terminate(new_status)
 
 
 class SyncArrival(AtomicBehavior):
