+++ srunner/scenarios/control_loss.py
@@ -16,9 +16,10 @@ import random
 
 import py_trees
 
-from srunner.scenariomanager.atomic_scenario_behavior import *
-from srunner.scenariomanager.atomic_scenario_criteria import *
-from srunner.scenarios.basic_scenario import *
+from srunner.scenariomanager.scenarioatomics.atomic_behaviors import *
+from srunner.scenariomanager.scenarioatomics.atomic_criteria import CollisionTest
+from srunner.scenariomanager.scenarioatomics.atomic_trigger_conditions import *
+from srunner.scenarios.basic_scenario import BasicScenario
 from srunner.tools.scenario_helper import *
 
 CONTROL_LOSS_SCENARIOS = [
