+++ srunner/scenarios/object_crash_intersection.py
@@ -13,10 +13,11 @@ import py_trees
 
 import carla
 
-from srunner.scenariomanager.atomic_scenario_behavior import *
-from srunner.scenariomanager.atomic_scenario_criteria import *
+from srunner.scenariomanager.scenarioatomics.atomic_behaviors import *
+from srunner.scenariomanager.scenarioatomics.atomic_criteria import *
+from srunner.scenariomanager.scenarioatomics.atomic_trigger_conditions import *
 from srunner.scenariomanager.timer import TimeOut
-from srunner.scenarios.basic_scenario import *
+from srunner.scenarios.basic_scenario import BasicScenario
 from srunner.tools.scenario_helper import *
 
 VEHICLE_TURNING_SCENARIOS = [
