+++ srunner/scenarios/other_leading_vehicle.py
@@ -21,10 +21,11 @@ import py_trees
 
 import carla
 
-from srunner.scenariomanager.atomic_scenario_behavior import *
-from srunner.scenariomanager.atomic_scenario_criteria import *
+from srunner.scenariomanager.scenarioatomics.atomic_behaviors import *
+from srunner.scenariomanager.scenarioatomics.atomic_criteria import *
+from srunner.scenariomanager.scenarioatomics.atomic_trigger_conditions import *
 from srunner.scenariomanager.carla_data_provider import CarlaDataProvider
-from srunner.scenarios.basic_scenario import *
+from srunner.scenarios.basic_scenario import BasicScenario
 from srunner.tools.scenario_helper import get_waypoint_in_distance
 
 OTHER_LEADING_VEHICLE_SCENARIOS = [
