+++ srunner/scenarios/signalized_junction_left_turn.py
@@ -16,9 +16,10 @@ import py_trees
 import carla
 from agents.navigation.local_planner import RoadOption
 
-from srunner.scenariomanager.atomic_scenario_behavior import *
-from srunner.scenariomanager.atomic_scenario_criteria import *
-from srunner.scenarios.basic_scenario import *
+from srunner.scenariomanager.scenarioatomics.atomic_behaviors import *
+from srunner.scenariomanager.scenarioatomics.atomic_criteria import *
+from srunner.scenariomanager.scenarioatomics.atomic_trigger_conditions import *
+from srunner.scenarios.basic_scenario import BasicScenario
 from srunner.tools.scenario_helper import *
